+++ daemon/algod/api/server/v2/test/helpers.go
@@ -252,7 +252,7 @@ func testingenv(t testing.TB, numAccounts, numTxs int, offlineAccounts bool) (*d
 		}
 	}
 
-	genesis[poolAddr] = basics.MakeAccountData(basics.NotParticipating, basics.MicroAlgos{Raw: 100000 * uint64(proto.RewardsRateRefreshInterval)})
+	genesis[poolAddr] = basics.MakeAccountData(basics.NotParticipating, basics.MicroAlgos{Raw: 100000 * proto.RewardsRateRefreshInterval})
 
 	bootstrap := data.MakeGenesisBalances(genesis, poolAddr, sinkAddr)
